[api-merge] Allow sorting API descriptions via `api-merge -s`.
`api-merge` processes API descriptions in argument list order, and
requires that API descriptions be in ascending order, meaning
this is good:

	api-merge.exe -o merged.xml api-1.xml api-2.xml api-3.xml

and this isn't good:

	api-merge.exe -o merged.xml api-3.xml api-2.xml api-1.xml

(Descriptions need to be in ascending order so that API additions and
movements can be properly tracked.)

This was fine in the Old World Order™: a make(1) variable ensured that
ordering was ascending:

	API_LEVELS = 1 2 3
	merged.xml:
		mono api-merge.xml -o merged.xml $(API_LEVELS:%=api-%.xml)

In the glorious new MSBuild-oriented order, we don't yet have an
equivalent to $(API_LEVELS), nor am I sure how to sanely represent
that (perhaps as an @(ApiLevels) ItemGroup?).

However, we still need a way to invoke api-merge.exe with files in the
correct order. There are thus two options:

 1. Rely on MSBuild file glob references to be sorted correctly:

	<ItemGroup>
	  <Profile Include="Profiles\*.xml.in" />
	</ItemGroup>

    and hope that @(Profile) is sorted correctly, or

 2. Improve api-merge.exe to sort the descriptions.

I'm not keen on (1) -- I don't see why files should be sorted from
MSBuild -- which leaves (2):

Add an `api-merge -s GLOB` option so that description sorting can be
controlled. If `-s GLOB` isn't provided, inputs are not sorted:

	# Bad order, descriptions are unsorted
	api-merge.exe -o merged.xml api-3.xml api-1.xml

	# Bad order, -s provided, descriptions are properly sorted
	api-merge.exe -o merged.xml -s 'api-*.xml' api-3.xmlm api-1.xml

Using `-s GLOB` thus removes any dependency on MSBuild (not) sorting
files, while still ensuring that API descriptions are processed in the
correct order.

GLOB sorting is...a bit of a hack. The GLOB is translated into a
regular expression, and each glob character ('*', '?') is turned into
a regular expression group. When sorting, each group is processed
in-turn, we attempt to parse as an int, and if the parse passes, we
sort by the int values.