[api-merge] Add `api-merge --last-description=FILE` parameter.
Consider the MSBuild-usage scenario implied in 11143d18:

	<ItemGroup>
		<Profile Include="Profiles\api-*.xml" />
	</ItemGroup>
	<Exec Command="api-merge.exe -o merged.xml -s 'api-*.xml' @(Profile->'%(Identity)', ' ')" />

Assuming the files Profiles/api-10.xml and Profiles/api-15.xml, the
above will execute the command:

	api-merge.exe -o merged.xml Profiles/api-10.xml Profiles/api-15.xml -s 'api-*.xml'

This seems fine...until you want to generate a merged XML description
for API-10, *not* API-15. At which point we're not emitting the
correct description, which is a problem.

To fix this, add `api-merge --last-description=NAME`, so that we can
control which is the last file to process. If NAME is found, then
API description processing will stop after encountering that file.
If NAME is *not* found, then all files will be processed.

This allows generating a merged API-10 desciption via:

	<Exec Command="api-merge.exe -o merged.xml -s 'api-*.xml' --last-description=Profiles/api-10.xml @(Profile->'%(Identity)', ' ')" />