[mono-runtimes] Fix _BuildRuntimes @Outputs checking.
The _BuildRuntimes target was executing unnecessarily:

	Target _BuildRuntimes needs to be built as input file
	'obj/Debug//armeabi-v7a/.stamp' is newer than output file
	'obj/host/mono/profiler/.libs/libmono-profiler-log.dylib

This would *always* be true, because the output files were never
touched once they were built, so e.g. when copying a build tree from
one machine to another (Dropbox!), the dependency check would be
"wrong".

Improve the _BuildRuntimes target to explicitly <Touch/> the output
files so that -- short of timezone changes (!) -- the outputs will be
forced to be newer than the inputs.