[Xamarin.Android.Build.Tasks] Fix the inclusion of .config files in the Apk. (#16)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=40804

There was an issue with including config files in the final apk when
using release builds. Firstly they were not even being picked up
because the CopyConfigFiles task was expecting a list of config files
not a list of assemblies which is what it was receiving.

Secondly the BuildApk task was not placing them in the assemblies
directory.

This commit fixes both of those issues.