Add Code formatting Policies for Xamarin Studio/MonoDevleop
This commit adds the MonoDevelopPolicies.xml file to the repo. This
will enforce the code formatting for the various projects in the
repo. Currently it is enforcing the standard Visual Studio formatting.
We will probably need to add the same for Visual Studio as well just
in case the end user overrides the defaults.