Fix Configuration for Xamarin.Android.sln
If you loaded the .sln in Xamarin studio it defaults to the
`Debug` configuration. As a result the Java.Interop project is
NOT built as part of that configuration. This is because it was
looking for a Mapping that did not exist in that project.

While it didn't stop the build, it did looks odd and might cause
confusion for users.

Other configurations have also been updated to make sure all the
projects are building the correct projects