[msbuild] Applications crashing with `System.DllNotFoundException: sqlite3` for projects that reference sqlite but don't use fastdev (#415) (#25)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=40804

There was an issue with the way the dll.config files were
extracted from the apk. They were including additional data
at the end of the xml because the files were not null
terminated.

This patch fixes the null termination for user included
config files and those included in the debug runtime.
It also adds method to extract the config file from the apk
(this was apparently missing too).