[android-toolchain] Fix the location of the android tools
The current directory structure for the android sdk does NOT
match the one that google creates. The tools are not in a
tools subdirectory when they should be. As a result things like
the sdkmanager do not run. It also causes problems for Xamarin Studio
since its looking for things in different places.

This commit fixes up the tools .zip that we download to make sure
the tools are extacted to the correct directory. This is because
the new zip file contains the tools directory within it rather than
just being a flat zip file. Also it is the recommended tools download
from google [1]

[1] http://developer.android.com/sdk/index.html#mac-tools