[Xamarin.Android.Build.Tasks] VS hangs/freezes when opening solutions containing Android Bindings Libraries that reference other libraries, during `GetAdditionalResourcesFromAssemblies.Execute()` (#28)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=40982

The Binding targets had the same issue as the normal Common
targets for android projects. Commit d300845 added a new
target to set $(DesignTimeBuild) to false if it was not
set. This way when Visual studio does an intellisense build
we do not attempt to download additional resources.

This commit makes the same changes to the Bindings targets.