Use HTTPS URLs for git submodules
Relative URLs for the submodules will make it harder to work with
various versions of Mono and Java.Interop checked out in parallel
with the proprietary parts of Xamarin.Android as both parties
would use the same location for the "master" repository but different
hashes or branches. This would lead to frequent rebuilding of both
submodules as developers switch between the OSS and proprietary
repositories.

Additionally, using a relative URL like previously means that we
have no information where the submodule was checked out from. Should
the repository referred to in the submodule URL be removed from
filesystem, user might have a problem recreating it (the same applies
to cloning the xamarin-android repository in which case to get the
submodules initialized one has to first check out Mono and Java.Interop
in the directory above xamarin-android - but there's no information
where to check them out from)

Furthermore, updating the submodule would require going first to
the "master" repository above xamarin-android, pulling the changes
there, going back to xamarin-android to update/init submodules.
HTTPs URLs fix the original issue of requiring that the users are
logged in to GitHub in order to clone xamarin-android and it avoids all
the pitfalls outlined above.