[Xamarin.Android.Build.Tasks] Fix source profiler library name.
The `_InstallRuntimes` target in
`build-tools/mono-runtimes/mono-runtimes.targets` copies over
`libmono-profiler-log.so` into
`bin/$(Configuration)/lib/xbuild/Xamarin/Android/lib/ABI/libmono-profiler-log.so`.

Unfortunately, the `<BuildApk/>` task looks for
`libmono-profiler-log.ABI.so`, which is a file that doesn't exist.

Result: If `BuildApk.AndroidEmbedProfilers` is not the empty string
and/or `BuildApk.Debug` is "true" (`$(AndroidIncludeDebugSymbols)` is
True, e.g. `$(DebugSymbols)` is True and `$(DebugType)` is Full), then
the build breaks:

	Error executing task BuildApk: Could not find file
	".../bin/Debug/lib/xbuild/Xamarin/Android/lib/armeabi-v7a/libmono-profiler-log.armeabi-v7a.so".

Fix the source filename so that profilers can be properly embedded.