Be more granular with SDK and NDK stamp files
Previously we used a single stamp file to detect
whether SDK/NDK need to be updated/recreate. This
worked fine if nobody touched the toolchain directory.
However, if any of the directories (sdk, ndk or anything in them)
were removed, build would *not* recreate the toolchain *unless*
the stamp file in the root folder would be deleted as well.

This patch creates a stamp file per component directory for all
the NDK and SDK components and thus the content will be restored
should the directory be removed.

It can be further improved, probably, by not removing the entire
directory prior to unzipping files as we do now but instead unpack
only those components that are missing.