Build mono if monodroid profile is missing
If the mono build was aborted for whatever reason and
the monodroid profile wasn't built completely (or at all)
the current targets will not fail because they won't see
any of the outputs missing and the _InstallBcl target will
happily accept empty set for @(_Assemblies).

We currently depend on a manually added full list of
the monodroid profile assemblies (sans Mono.Data.Sqlite which
will is built separately elsewhere and will not be built
as part of the profile eventually), but we will implement
a way to dynamically obtain the list of assemblies from mono.