[android-toolchains, mono-runtimes] Build armeabi, arm64-v8a, x86, x86_64
Commit 38dbfcaf mentions that commercial Xamarin.Android 6.0
provides Mono for five architectures, but "[i]n the interest of
expediency" only adds support to build *one* architecture:
armeabi-v7a (32-bit ARM v7).

It's time to fix that: add build system support for armeabi,
arm64-v8a, x86, and x86_64.

*However*, it takes *time* to build all those ABIs: on a 2013 6-core
Mac Pro, it takes ~29 minutes to build all five of those ABIs plus the
"host" ABI (for BCL assemblies), which is presumably 29 minutes that
very few people want to spend, and will be even longer in a variety of
build environments (virtual machines, slower hardware, etc.).

Which means we don't want to require that they all be built.

To better support this, add a new `$(AndroidSupportedAbis)` MSBuild
property which contains a comma-separated list of ABIs to support.
This allows manually overriding the ABIs on the command-line:

	# build everything!
	$ xbuild /p:AndroidSupportedAbis=armeabi,armeabi-v7a,arm64-v8a,x86,x86_64

or setting a value within `Configuration.Override.props`:

	<PropertyGroup>
	  <!-- only build x86 -->
	  <AndroidSupportedAbis>x86,x86_64</AndroidSupportedAbis>
	</PropertyGroup>

The *default* continues to be just armeabi-v7a.