[build-tools] Update library outputs to not include <abi> suffix
The native libraries are output into an <abi> specific folder in

	lib/xbuild/Xamarin/Android/libs

As a result we really do not need to have an <abi> suffix as part
of the native library name. In additon the libmonosgen library was
having -unstripped appended to the end of the library so it was

	libmonosgen-2.0.so-unstripped

this looks a bit weird since it replaces the normal .so extension.
This has been removed.