[generator] Emit correct XPath references
Comments are inserted into the generated C# code to identify the
emitted member with the "source" XML declaration, to permit later
fixups via metadata without the drudgery of manually determining the
XPath references.

Unfortunately, when a class implements an interface, the generated
XPath comment for the interface method within the class would provide
an XPath expression "as if" the method were declared in the class,
which was not the case:

	// Metadata.xml XPath method reference: path="/api/package[@name='java.util']/class[@name='AbstractCollection']/method[@name='spliterator' and count(parameter)=0]"
	[Register ("spliterator", "()Ljava/util/Spliterator;", "GetSpliteratorHandler", ApiSince = 24)]
	public virtual unsafe Java.Util.ISpliterator Spliterator ()

Above, the spliterator() method is *actually* declared in
java.util.Spliterator<T>, but the comment implies that the generated
method is coming from java.util.AbstractCollection<T>. The comment is
thus wrong, as AbstractCollection.spliterator() doesn't exist within
the API XML (it's implied by the implemented interfaces).

To fix this, use the Method's DeclaringType in the XPath comment:

	// Metadata.xml XPath method reference: path="/api/package[@name='java.util']/class[@name='Spliterator']/method[@name='spliterator' and count(parameter)=0]"