[jcw-gen] Generate Java Callable Wrappers in parallel.
Generate Java Callable Wrappers across multiple tasks.

A note about the tests: when I first added jcw-gen I tried to use
Tasks so I could get parallel JCW generation, but some of the output
was unexpected, in that it would have the wrong package name.

This is why the JavaCallableWrapperGeneratorTests.GenerateInParallel()
test was added: to try to reproduce the error that I recalled seeing.

Doing this found two unexpected bugs in `JniType.ToJniName()` (yay):

1. I was getting a `NullReferenceException` from
    `JniType.IsNonStaticInnerClass()`, because in
    `JniType.ToJniName<T>()` there is a call to
    `JniType.IsNonStaticInnerClass()` using the result of
    `as TypeDefinition`, but `IsNonStaticInnerClass()` never checked
    for `null`. Oops.

2. The reflection-based `JniType.ToJniName(Type)` overload didn't
    check for `Android.App.InstrumentationAttribute` providing an
    overriding `InstrumentationAttribute.Name` property.