[Configuration] Use `:`, not `,`, as ABI separator.
Update `$(AndroidSupportedAbis)` and
`$(AndroidSupportedAbisForConditionalChecks)` to use `:` as the ABI
separator, not `,`.

`;` can't be used because xbuild and MSBuild don't like `;` as a
property value on the command-line.

Turns out, MSBuild doesn't like `,` within property values on the
command-line either, because it allows multiple MSBuild properties to
be specified via one `/p:` use:

	$ msbuild -help
	...
	  /property:<n>=<v>  Set or override these project-level properties. <n> is
	                     the property name, and <v> is the property value. Use a
	                     semicolon or a comma to separate multiple properties, or
	                     specify each property separately. (Short form: /p)
	                     Example:
	                       /property:WarningLevel=2;OutDir=bin\Debug\

This means that it's not possible to set `$(AndroidSupportedAbis)` to
e.g. `host-Darwin,armeabi-v8a` on the command-line *with MSBuild*.
(It is with xbuild, but this is arguably an xbuild compatibility bug!)

Since we want to be able to easily override the
`$(AndroidSupportedAbis)` value on the command-line for testing,
change the ABI separator character to `:` which is supported on both
xbuild and MSBuild:

	$ xbuild /p:AndroidSupportedAbis=host-Darwin:armeabi-v7a
	# works!