[Xamarin.Android.Build.Tasks] Fix the creation of Library Resource Archive
The commit 4ec06ac9 broke the way resource archives are creating. Firstly
it was writting an empty zip file because we were not closing the zip after
adding the files BEFORE writing the stream to disk. So it had not been
flushed.

Secondaly the directory structure within the zip itself was incorrect. It
was adding files using the full path rather than the internal path that
we needed in the zip.

This commit fixes all of that.