Add a thorough clean make target
`make distclean` will not only run `make clean` but it will also
run `git clean -xdff` for the main project as well as the submodules
thus giving you a pristine tree.

Note that it will remove *all* of the files not in the git index
as well as all the build artifacts! Use with care!