[Xamarin.Android.Build.Tasks] Fix issues with Library Resource Extraction (#69)
Since the switch to System.IO.Compression we have a few
issues when dealing with resource extraction.

1) It was not ignoring __MACOSX or .DS_Store during extraction.
2) It was not correctly detecting duplicate zip entries.

This commit updates the code to correctly filter __MACOSX and
.DS_Store entries. It also changes the way we check if an entry
exists in the zip already. And finally it updates
ResolveLibraryProjectImports to use the Files.ExtractAll method.
This is because the standard ZipFile.ExtractToDirectory has no
way of filtering the output (so it includes __MACOSX and .DS_Store).