Make Mono.Data.Sqlite smaller
Don't include all the shared sqlite3 libraries but only the ones that
correspond to the ABIs this particular build supports.

In order to make it possible a new property is introduced in
Configuration.props - AndroidSupportedDeviceAbis - which contains *just*
the target Android device ABIs without the `host-*` one. The extra
property is necessary because we don't want to include anything `host-*`
in the Android assemblies and, at the same time, we cannot use the
AndroidSupportedAbisForConditionalChecks property since it separates its
components with `:` to make conditionals easier and thus it would require
extra processing wherever we need just the list of the enabled ABIs.

Also, don't build the native sqlite3 library for all the supported ABIs.
Instead, build it just for the enabled ones - it saves quite a bit of
time.