[Mono.Posix] Sign with `mono.pub`
Sign `Mono.Posix.dll` with `mono.pub`, not with `product.snk`.

This is required because the existing Xamarin.Android SDK signs
`Mono.Posix.dll` with `mono.pub`, so the same signing key needs to be
used so that existing assembly references are not broken because a
new/different signing key is used.