[build] Add $(JavaInteropSourceDirectory) MSBuild property
Allow the Java.Interop checkout directory to be specified by
overriding the `$(JavaInteropSourceDirectory)` MSBuild property.

Project files should use the `$(JavaInteropFullPath)` MSBuild
property, which is computed as `Path.GetFullPath()` for the
`$(JavaInteropSourceDirectory)` value.

Normally `$(JavaInteropSourceDirectory)` wouldn't need to be
overridden; the current use case is to allow a CI-like environment
which grabs the latest commit of every referenced module to make sure
they all work together. (Note: such a "CI-like environment"
DOES NOT (yet) EXIST. This commit is to help *permit* such a thing.)